requirements.txt
langchain
psycopg2-binary
python-dotenv
openai
